Refactor: Remove abstract class, merge into one.
